Add files via upload
Project: Modelling semantic, machine-learning-based text analysis. Supervisor: prof. dr. hab. inż. Grzegorz Nalepa.

Semantic analysis of a text is a complex function. Selected filters, machine learning techniques, text complexity and analyzed language all influence the effectiveness of the model. The rationale for the project was to compare the efficacy of different machine-learning methods and factors for both the English and Polish language.